requirements.txt
pytest>=7.0
python-dotenv>=1.0
kafka-python==2.0.2